model and texture update

+++ src/interface/renderer/texture/texture_location.rs
@@ -1,11 +1,11 @@
 #[derive(Clone, PartialEq)]
 pub struct TextureLocation {
-    pub atlas_index: u8,
-    pub layer_index: u8,
+    pub atlas_index: usize,
+    pub layer_index: usize,
 }
 
 impl TextureLocation {
-    pub fn new(atlas_index: u8, layer_index: u8) -> Self {
+    pub fn new(atlas_index: usize, layer_index: usize) -> Self {
         Self {
             atlas_index,
             layer_index,
